refactor(ios): extract and polish reader views into its own file
- extract views from reader iOSReaderView and iOSReaderTabView
- add previews to views with custom values.
- improve iOSReaderTabView books tab visibility.
- disable fatalError on DatabaseManager when using preview environment.

+++ Source/Managers/Database/DatabaseManager.swift
@@ -101,6 +101,9 @@ class DatabaseManager {
                 comment: ""
             )
         )
+        if ProcessInfo.processInfo.environment["XCODE_RUNNING_FOR_PREVIEWS"] == "1" {
+            return
+        }
         #if os(macOS)
         NSApp.terminate(nil)
         #else
